Added Player Movement (only 4 directions)

+++ Entities/Hittable.cs
@@ -1,11 +1,9 @@
 public abstract class Hittable : Entity
 {
-    protected Hittable(string path) : base(path)
-    {
-    }
+    protected Hittable(string path) : base(path) { }
 
-    public int X { get; set; }
-    public int Y { get; set; }
+    public float X { get; set; }
+    public float Y { get; set; }
     public bool Colided { get; set; }
 
     public abstract bool Colision(Hittable hittable);
